+++ myokit/tests/test_config.py
@@ -14,15 +14,14 @@ import unittest
 
 import myokit
 
-from myokit.tests import TemporaryDirectory
+from myokit.tests import TemporaryDirectory, WarningCollector
 
 
-# Empty config
-config1 = """
-"""
+debug = True
 
-# Config with lots of settings
-config2 = """
+
+# Simple test
+config_basic = """
 [myokit]
 [time]
 date_format = TEST_DATE_FORMAT
@@ -36,8 +35,16 @@ lib = five;six
 inc = three;eight
 """
 
+# Config with extra bits
+config_extra = """
+[extra]
+setting = 12
+sotting=2
+str=String value
+"""
+
 # Config with empty paths and spaces
-config_empties_1 = """
+config_paths_1 = """
 [sundials]
 lib = five
 inc = three;four;
@@ -47,7 +54,7 @@ inc =
 """
 
 # Config with empty paths and " ;", which in Python 2 is ignored
-config_empties_2 = """
+config_paths_2 = """
 [sundials]
 lib = five ; six
 inc = three;four;
@@ -56,58 +63,11 @@ lib = one ;;   two point five ;three;
 inc =
 """
 
-# Qt options
-config_pyside = """
-[gui]
-backend=pyside
-"""
-
-config_pyside2 = """
-[gui]
-backend=pyside2
-"""
-
-config_pyqt4 = """
-[gui]
-backend=pyqt4
-"""
-
-config_pyqt5 = """
-[gui]
-backend=pyqt5
-"""
-
-# Config with extra bits
-config3 = """
-[extra]
-setting = 12
-sotting=2
-str=String value
-"""
-
 
 class TestConfig(unittest.TestCase):
 
-    def test_create(self):
-        # Test if the `_create` method works.
-
-        # Import hidden _config module
-        path = sys.path
-        try:
-            sys.path.append(myokit.DIR_MYOKIT)
-            import _config as config
-        finally:
-            sys.path = path
-
-        # Test _create
-        with TemporaryDirectory() as d:
-            filename = d.path('test.ini')
-            config._create(filename)
-            self.assertTrue(os.path.isfile(filename))
-        self.assertFalse(os.path.isfile(filename))
-
-    def test_load_read(self):
-        # Test if the `_load` method works, when a config file exists.
+    @classmethod
+    def setUpClass(cls):
 
         # Import hidden _config module
         path = sys.path
@@ -118,157 +78,286 @@ class TestConfig(unittest.TestCase):
             sys.path = path
 
         # Back-up current settings
-        date_format = myokit.DATE_FORMAT
-        time_format = myokit.TIME_FORMAT
-        force_pyside = myokit.FORCE_PYSIDE
-        force_pyside2 = myokit.FORCE_PYSIDE2
-        force_pyqt4 = myokit.FORCE_PYQT4
-        force_pyqt5 = myokit.FORCE_PYQT5
-        sundials_lib = myokit.SUNDIALS_LIB
-        sundials_inc = myokit.SUNDIALS_INC
-        opencl_lib = myokit.OPENCL_LIB
-        opencl_inc = myokit.OPENCL_INC
+        cls._compat_no_capture = myokit.COMPAT_NO_CAPTURE
+        cls._compat_no_fd_capture = myokit.COMPAT_NO_FD_CAPTURE
+        cls._date_format = myokit.DATE_FORMAT
+        cls._time_format = myokit.TIME_FORMAT
+        cls._force_pyside = myokit.FORCE_PYSIDE
+        cls._force_pyside2 = myokit.FORCE_PYSIDE2
+        cls._force_pyqt4 = myokit.FORCE_PYQT4
+        cls._force_pyqt5 = myokit.FORCE_PYQT5
+        cls._sundials_lib = myokit.SUNDIALS_LIB
+        cls._sundials_inc = myokit.SUNDIALS_INC
+        cls._opencl_lib = myokit.OPENCL_LIB
+        cls._opencl_inc = myokit.OPENCL_INC
 
-        # Change myokit config dir temporarily
-        path = myokit.DIR_USER
-        try:
-            with TemporaryDirectory() as d:
-                myokit.DIR_USER = d.path()
+        # Temporarily change myokit config dir
+        cls._dir_user = myokit.DIR_USER
+        cls._temp_dir = TemporaryDirectory()
+        cls._temp_dir.__enter__()
+        myokit.DIR_USER = cls._temp_dir.path()
 
-                # Simple test
-                with open(d.path('myokit.ini'), 'w') as f:
-                    f.write(config1)
-                config._load()
+        # Config module
+        cls._config_module = config
 
-                # Full values, PySide gui
-                myokit.SUNDIALS_LIB = []
-                myokit.SUNDIALS_INC = []
-                myokit.OPENCL_LIB = []
-                myokit.OPENCL_INC = []
-                myokit.FORCE_PYSIDE = myokit.FORCE_PYSIDE2 = False
-                myokit.FORCE_PYQT4 = myokit.FORCE_PYQT5 = False
-                with open(d.path('myokit.ini'), 'w') as f:
-                    f.write(config2)
-                config._load()
-                self.assertEqual(myokit.DATE_FORMAT, 'TEST_DATE_FORMAT')
-                self.assertEqual(myokit.TIME_FORMAT, 'TEST_TIME_FORMAT')
-                self.assertFalse(myokit.FORCE_PYSIDE)
-                self.assertFalse(myokit.FORCE_PYSIDE2)
-                self.assertFalse(myokit.FORCE_PYQT4)
-                self.assertFalse(myokit.FORCE_PYQT5)
-                self.assertEqual(myokit.SUNDIALS_LIB, ['one', 'two'])
-                self.assertEqual(myokit.SUNDIALS_INC, ['three', 'four'])
-                self.assertEqual(myokit.OPENCL_LIB, ['five', 'six'])
-                self.assertEqual(myokit.OPENCL_INC, ['three', 'eight'])
+        if debug:
+            print('setUpClass completed')
 
-                # Lists of paths should be filtered for empty values and
-                # trimmed
-                myokit.SUNDIALS_LIB = []
-                myokit.SUNDIALS_INC = []
-                myokit.OPENCL_LIB = []
-                myokit.OPENCL_INC = []
-                with open(d.path('myokit.ini'), 'w') as f:
-                    f.write(config_empties_1)
-                config._load()
-                self.assertEqual(myokit.SUNDIALS_LIB, ['five'])
-                self.assertEqual(myokit.SUNDIALS_INC, ['three', 'four'])
-                self.assertEqual(
-                    myokit.OPENCL_LIB,
-                    ['one', 'two point five', 'three'])
-                self.assertEqual(myokit.OPENCL_INC, [])
+    @classmethod
+    def tearDownClass(cls):
 
-                # Even if the list contains " ;", which Python 2's config
-                # parser treats as a comment
-                myokit.SUNDIALS_LIB = []
-                myokit.SUNDIALS_INC = []
-                myokit.OPENCL_LIB = []
-                myokit.OPENCL_INC = []
-                with open(d.path('myokit.ini'), 'w') as f:
-                    f.write(config_empties_2)
-                if sys.hexversion < 0x03020000:
-                    self.assertRaises(ImportError, config._load)
-                else:
-                    config._load()
-                    self.assertEqual(myokit.SUNDIALS_LIB, ['five', 'six'])
-                    self.assertEqual(myokit.SUNDIALS_INC, ['three', 'four'])
-                    self.assertEqual(
-                        myokit.OPENCL_LIB,
-                        ['one', 'two point five', 'three'])
-                    self.assertEqual(myokit.OPENCL_INC, [])
+        # Reset path
+        myokit.DIR_USER = cls._dir_user
 
-                # Qt gui options
-                with open(d.path('myokit.ini'), 'w') as f:
-                    f.write(config_pyqt4)
-                config._load()
-                self.assertFalse(myokit.FORCE_PYSIDE)
-                self.assertFalse(myokit.FORCE_PYSIDE2)
-                self.assertTrue(myokit.FORCE_PYQT4)
-                self.assertFalse(myokit.FORCE_PYQT5)
+        # Reset data and time
+        myokit.COMPAT_NO_CAPTURE = cls._compat_no_capture
+        myokit.COMPAT_NO_FD_CAPTURE = cls._compat_no_fd_capture
+        myokit.DATE_FORMAT = cls._date_format
+        myokit.TIME_FORMAT = cls._time_format
+        myokit.FORCE_PYSIDE = cls._force_pyside
+        myokit.FORCE_PYSIDE2 = cls._force_pyside2
+        myokit.FORCE_PYQT4 = cls._force_pyqt4
+        myokit.FORCE_PYQT5 = cls._force_pyqt5
+        myokit.SUNDIALS_LIB = cls._sundials_lib
+        myokit.SUNDIALS_INC = cls._sundials_inc
+        myokit.OPENCL_LIB = cls._opencl_lib
+        myokit.OPENCL_INC = cls._opencl_inc
 
-                with open(d.path('myokit.ini'), 'w') as f:
-                    f.write(config_pyqt5)
-                config._load()
-                self.assertFalse(myokit.FORCE_PYSIDE)
-                self.assertFalse(myokit.FORCE_PYSIDE2)
-                self.assertFalse(myokit.FORCE_PYQT4)
-                self.assertTrue(myokit.FORCE_PYQT5)
+        # Reload local settings
+        cls._config_module._load()
 
-                with open(d.path('myokit.ini'), 'w') as f:
-                    f.write(config_pyside)
-                config._load()
-                self.assertTrue(myokit.FORCE_PYSIDE)
-                self.assertFalse(myokit.FORCE_PYSIDE2)
-                self.assertFalse(myokit.FORCE_PYQT4)
-                self.assertFalse(myokit.FORCE_PYQT5)
+        # Sanity check
+        assert myokit.COMPAT_NO_CAPTURE == cls._compat_no_capture
+        assert myokit.COMPAT_NO_FD_CAPTURE == cls._compat_no_fd_capture
+        assert myokit.DATE_FORMAT != 'TEST_DATE_FORMAT'
+        assert myokit.TIME_FORMAT != 'TEST_TIME_FORMAT'
+        assert myokit.FORCE_PYSIDE == cls._force_pyside
+        assert myokit.FORCE_PYSIDE2 == cls._force_pyside2
+        assert myokit.FORCE_PYQT4 == cls._force_pyqt4
+        assert myokit.FORCE_PYQT5 == cls._force_pyqt5
+        assert myokit.SUNDIALS_LIB != ['one', 'two']
+        assert myokit.SUNDIALS_INC != ['three', 'four']
+        assert myokit.OPENCL_LIB != ['five', 'six']
+        assert myokit.OPENCL_INC != ['three', 'eight']
 
-                with open(d.path('myokit.ini'), 'w') as f:
-                    f.write(config_pyside2)
-                config._load()
-                self.assertFalse(myokit.FORCE_PYSIDE)
-                self.assertTrue(myokit.FORCE_PYSIDE2)
-                self.assertFalse(myokit.FORCE_PYQT4)
-                self.assertFalse(myokit.FORCE_PYQT5)
+        if debug:
+            print('tearDownClass completed')
 
-                # Odd ini file
-                with open(d.path('myokit.ini'), 'w') as f:
-                    f.write(config3)
-                config._load()
+    def test_create(self):
+        # Test if the `_create` method works.
 
-                # No ini file (calls _create())
-                os.remove(d.path('myokit.ini'))
-                config._load()
+        # Test _create
+        with TemporaryDirectory() as d:
+            filename = d.path('test.ini')
+            self._config_module._create(filename)
+            self.assertTrue(os.path.isfile(filename))
+        self.assertFalse(os.path.isfile(filename))
 
-        finally:
-            # Reset path
-            myokit.DIR_USER = path
+        # No ini file (calls _create())
+        path = self._temp_dir.path('myokit.ini')
+        if os.path.exists(path):
+            os.remove(path)
+        self._config_module._load()
+        self.assertTrue(os.path.isfile(path))
+        os.remove(path)
+        self.assertFalse(os.path.isfile(path))
 
-            # Reset data and time
-            myokit.DATE_FORMAT = date_format
-            myokit.TIME_FORMAT = time_format
-            myokit.FORCE_PYSIDE = force_pyside
-            myokit.FORCE_PYSIDE2 = force_pyside2
-            myokit.FORCE_PYQT4 = force_pyqt4
-            myokit.FORCE_PYQT5 = force_pyqt5
-            myokit.SUNDIALS_LIB = sundials_lib
-            myokit.SUNDIALS_INC = sundials_inc
-            myokit.OPENCL_LIB = opencl_lib
-            myokit.OPENCL_INC = opencl_inc
+    def test_load_basic(self):
+        # Tests basic entries are loaded.
+        myokit.SUNDIALS_LIB = []
+        myokit.SUNDIALS_INC = []
+        myokit.OPENCL_LIB = []
+        myokit.OPENCL_INC = []
+        myokit.FORCE_PYSIDE = myokit.FORCE_PYSIDE2 = False
+        myokit.FORCE_PYQT4 = myokit.FORCE_PYQT5 = False
+        with open(self._temp_dir.path('myokit.ini'), 'w') as f:
+            f.write(config_basic)
+        self._config_module._load()
+        self.assertEqual(myokit.DATE_FORMAT, 'TEST_DATE_FORMAT')
+        self.assertEqual(myokit.TIME_FORMAT, 'TEST_TIME_FORMAT')
+        self.assertFalse(myokit.FORCE_PYSIDE)
+        self.assertFalse(myokit.FORCE_PYSIDE2)
+        self.assertFalse(myokit.FORCE_PYQT4)
+        self.assertFalse(myokit.FORCE_PYQT5)
+        self.assertEqual(myokit.SUNDIALS_LIB, ['one', 'two'])
+        self.assertEqual(myokit.SUNDIALS_INC, ['three', 'four'])
+        self.assertEqual(myokit.OPENCL_LIB, ['five', 'six'])
+        self.assertEqual(myokit.OPENCL_INC, ['three', 'eight'])
 
-            # Reload local settings
-            config._load()
+        # Unknown settings are ignored
+        with open(self._temp_dir.path('myokit.ini'), 'w') as f:
+            f.write(config_extra)
+        self._config_module._load()
 
-            # Sanity check
-            self.assertNotEqual(myokit.DATE_FORMAT, 'TEST_DATE_FORMAT')
-            self.assertNotEqual(myokit.TIME_FORMAT, 'TEST_TIME_FORMAT')
-            self.assertEqual(myokit.FORCE_PYSIDE, force_pyside)
-            self.assertEqual(myokit.FORCE_PYSIDE2, force_pyside2)
-            self.assertEqual(myokit.FORCE_PYQT4, force_pyqt4)
-            self.assertEqual(myokit.FORCE_PYQT5, force_pyqt5)
-            self.assertNotEqual(myokit.SUNDIALS_LIB, ['one', 'two'])
-            self.assertNotEqual(myokit.SUNDIALS_INC, ['three', 'four'])
-            self.assertNotEqual(myokit.OPENCL_LIB, ['five', 'six'])
-            self.assertNotEqual(myokit.OPENCL_INC, ['three', 'eight'])
+    def test_load_empty(self):
+        # Test loading with an empty file.
+
+        with open(self._temp_dir.path('myokit.ini'), 'w') as f:
+            f.write('')
+        self._config_module._load()
+
+    def test_load_paths(self):
+        # Test loading of sundials and opencl paths
+
+        # Lists of paths should be filtered for empty values and trimmed
+        myokit.SUNDIALS_LIB = []
+        myokit.SUNDIALS_INC = []
+        myokit.OPENCL_LIB = []
+        myokit.OPENCL_INC = []
+        with open(self._temp_dir.path('myokit.ini'), 'w') as f:
+            f.write(config_paths_1)
+        self._config_module._load()
+        self.assertEqual(myokit.SUNDIALS_LIB, ['five'])
+        self.assertEqual(myokit.SUNDIALS_INC, ['three', 'four'])
+        self.assertEqual(
+            myokit.OPENCL_LIB,
+            ['one', 'two point five', 'three'])
+        self.assertEqual(myokit.OPENCL_INC, [])
+
+        # Even if the list contains " ;", which Python 2's config parser treats
+        # as a comment
+        myokit.SUNDIALS_LIB = []
+        myokit.SUNDIALS_INC = []
+        myokit.OPENCL_LIB = []
+        myokit.OPENCL_INC = []
+        with open(self._temp_dir.path('myokit.ini'), 'w') as f:
+            f.write(config_paths_2)
+        if sys.hexversion < 0x03020000:
+            self.assertRaises(ImportError, self._config_module._load)
+        else:
+            self._config_module._load()
+            self.assertEqual(myokit.SUNDIALS_LIB, ['five', 'six'])
+            self.assertEqual(myokit.SUNDIALS_INC, ['three', 'four'])
+            self.assertEqual(
+                myokit.OPENCL_LIB,
+                ['one', 'two point five', 'three'])
+            self.assertEqual(myokit.OPENCL_INC, [])
+
+    def test_load_gui_backend(self):
+        # Tests setting the GUI back end
+
+        # Qt gui options
+        with open(self._temp_dir.path('myokit.ini'), 'w') as f:
+            f.write('[gui]\nbackend=pyqt4\n')
+        self._config_module._load()
+        self.assertFalse(myokit.FORCE_PYSIDE)
+        self.assertFalse(myokit.FORCE_PYSIDE2)
+        self.assertTrue(myokit.FORCE_PYQT4)
+        self.assertFalse(myokit.FORCE_PYQT5)
+
+        with open(self._temp_dir.path('myokit.ini'), 'w') as f:
+            f.write('[gui]\nbackend=pyqt5\n')
+        self._config_module._load()
+        self.assertFalse(myokit.FORCE_PYSIDE)
+        self.assertFalse(myokit.FORCE_PYSIDE2)
+        self.assertFalse(myokit.FORCE_PYQT4)
+        self.assertTrue(myokit.FORCE_PYQT5)
+
+        with open(self._temp_dir.path('myokit.ini'), 'w') as f:
+            f.write('[gui]\nbackend=pyside\n')
+        self._config_module._load()
+        self.assertTrue(myokit.FORCE_PYSIDE)
+        self.assertFalse(myokit.FORCE_PYSIDE2)
+        self.assertFalse(myokit.FORCE_PYQT4)
+        self.assertFalse(myokit.FORCE_PYQT5)
+
+        with open(self._temp_dir.path('myokit.ini'), 'w') as f:
+            f.write('[gui]\nbackend=pyside2\n')
+        self._config_module._load()
+        self.assertFalse(myokit.FORCE_PYSIDE)
+        self.assertTrue(myokit.FORCE_PYSIDE2)
+        self.assertFalse(myokit.FORCE_PYQT4)
+        self.assertFalse(myokit.FORCE_PYQT5)
+
+        # Empty has no effect
+        with open(self._temp_dir.path('myokit.ini'), 'w') as f:
+            f.write('[gui]\nbackend=\n')
+        with WarningCollector() as c:
+            self._config_module._load()
+        self.assertEqual(c.text(), '')
+        self.assertFalse(myokit.FORCE_PYSIDE)
+        self.assertTrue(myokit.FORCE_PYSIDE2)
+        self.assertFalse(myokit.FORCE_PYQT4)
+        self.assertFalse(myokit.FORCE_PYQT5)
+
+        # Invalid raises warning
+        with open(self._temp_dir.path('myokit.ini'), 'w') as f:
+            f.write('[gui]\nbackend=lalala\n')
+        with WarningCollector() as c:
+            self._config_module._load()
+        self.assertIn('Expected values for backend are', c.text())
+        self.assertTrue(myokit.FORCE_PYSIDE2)
+
+    def test_load_compat_no_capture(self):
+        # Tests loading compatibility option: no_capture
+
+        myokit.COMPAT_NO_CAPTURE = False
+        with open(self._temp_dir.path('myokit.ini'), 'w') as f:
+            f.write('[compatibility]\nno_capture = True\n')
+        self._config_module._load()
+        self.assertTrue(myokit.COMPAT_NO_CAPTURE)
+
+        with open(self._temp_dir.path('myokit.ini'), 'w') as f:
+            f.write('[compatibility]\nno_capture = False\n')
+        self._config_module._load()
+        self.assertFalse(myokit.COMPAT_NO_CAPTURE)
+
+        # Empty has no effect
+        with open(self._temp_dir.path('myokit.ini'), 'w') as f:
+            f.write('[compatibility]\nno_capture = \n')
+        with WarningCollector() as c:
+            self._config_module._load()
+        self.assertEqual(c.text(), '')
+        self.assertFalse(myokit.COMPAT_NO_CAPTURE)
+        myokit.COMPAT_NO_CAPTURE = True
+        with WarningCollector() as c:
+            self._config_module._load()
+        self.assertEqual(c.text(), '')
+        self.assertTrue(myokit.COMPAT_NO_CAPTURE)
+
+        # Invalid raises warning
+        with open(self._temp_dir.path('myokit.ini'), 'w') as f:
+            f.write('[compatibility]\nno_capture = hiya\n')
+        with WarningCollector() as c:
+            self._config_module._load()
+        self.assertTrue(myokit.COMPAT_NO_CAPTURE)
+        self.assertIn('Expected values for no_capture are', c.text())
+
+    def test_load_compat_no_capture_fd(self):
+        # Tests loading compatibility option: no_fd_capture
+
+        myokit.COMPAT_NO_FD_CAPTURE = False
+        with open(self._temp_dir.path('myokit.ini'), 'w') as f:
+            f.write('[compatibility]\nno_fd_capture = True\n')
+        self._config_module._load()
+        self.assertTrue(myokit.COMPAT_NO_FD_CAPTURE)
+
+        with open(self._temp_dir.path('myokit.ini'), 'w') as f:
+            f.write('[compatibility]\nno_fd_capture = False\n')
+        self._config_module._load()
+        self.assertFalse(myokit.COMPAT_NO_FD_CAPTURE)
+
+        # Empty has no effect
+        with open(self._temp_dir.path('myokit.ini'), 'w') as f:
+            f.write('[compatibility]\nno_fd_capture = \n')
+        with WarningCollector() as c:
+            self._config_module._load()
+        self.assertEqual(c.text(), '')
+        self.assertFalse(myokit.COMPAT_NO_FD_CAPTURE)
+        myokit.COMPAT_NO_FD_CAPTURE = True
+        with WarningCollector() as c:
+            self._config_module._load()
+        self.assertEqual(c.text(), '')
+        self.assertTrue(myokit.COMPAT_NO_FD_CAPTURE)
+
+        # Invalid raises warning
+        with open(self._temp_dir.path('myokit.ini'), 'w') as f:
+            f.write('[compatibility]\nno_fd_capture = hiya\n')
+        with WarningCollector() as c:
+            self._config_module._load()
+        self.assertTrue(myokit.COMPAT_NO_FD_CAPTURE)
+        self.assertIn('Expected values for no_fd_capture are', c.text())
 
 
 if __name__ == '__main__':
-    unittest.main()
+    unittest.main(verbosity=2)